7c7c3ca4 Add missing numpy functions previously available in ParaView 4.1.
f54b5b9f Merge topic 'enh_xdmf3_writer' into master
9f0d0bae Merge topic 'fix_xdmf2_writer' into master
fa2aef48 Merge topic 'update-hdf5' into master
acb40959 Enhance Xdmf3 writer to save block names.
2cc28ea5 Fix Xdmf2 writer: array type was not respected. Also save block name.
96db5108 Merge topic 'plotbar_improvement' into master
f7bf6158 Merge topic 'qt-warning-take3' into master
bed938eb Merge topic 'enum-comma-warning' into master
18f678ac Fix warning about comma at end of enum list.
91dcdaa6 Another attempt to silence warning from Qt header
4bcf0dc2 Prefix BUILD_STATIC_EXECS with HDF5 and make it advanced.
1dc7f42b Merge topic 'sample_distance_issue' into master
aab57b1d Merge topic 'fix_xdmf3_dll_export' into master
cac3a404 Merge topic 'type-punned-pointer' into master
1dcc3d96 Add functions to get data bounds and number of plotted bars.
...

+++ IO/Xdmf3/vtkXdmf3Reader.cxx
@@ -278,7 +278,7 @@ private:
   void Init(const char *filename, bool AsTime)
   {
     vtkTimerLog::MarkStartEvent("X3R::Init");
-    unsigned int idx = this->FileNames.size();
+    unsigned int idx = static_cast<unsigned int>(this->FileNames.size());
 
     this->Reader = XdmfReader::New();
